Code cleaning
Before moving on next step I just cleaned the code and make it more
understandable with renaming some components and variables.
